updated Unit types

+++ src/Components/PM/InprogressTable.jsx
@@ -299,7 +299,7 @@ const InprogressTable = ({ beneficiaries, value, setBeneficiaries, isReject, set
             accountNumber: parseInt(changedData.accountNumber, 10),
             revisedRatePerUnit: parseFloat(changedData.revisedRatePerUnit),
             benContribution: parseFloat(changedData.currentBeneficiaryContribution),
-            achievementUnit: parseInt(changedData.achievementUnit, 10),
+            achievementUnit: parseFloat(changedData.achievementUnit, 10),
             currentCost: parseFloat(changedData.currentCost),
             procurementCheck: changedData.procurementCheck === true,
             ...(isReject && { remark: changedData.remarks }),
